+++ README.md
@@ -103,7 +103,7 @@ Use `AsyncImednetSDK` when working with asyncio:
 
 ```python
 import os
-from imednet.async_sdk import AsyncImednetSDK
+from imednet.sdk import AsyncImednetSDK
 
 async def main():
     study_key = os.getenv("IMEDNET_STUDY_KEY", "your_study_key_here")
