+++ tests/live/test_endpoints_async_live.py
@@ -3,8 +3,7 @@ from typing import AsyncIterator, Iterator
 
 import pytest
 import pytest_asyncio
-from imednet.async_sdk import AsyncImednetSDK
-from imednet.sdk import ImednetSDK
+from imednet.sdk import AsyncImednetSDK, ImednetSDK
 
 API_KEY = os.getenv("IMEDNET_API_KEY")
 SECURITY_KEY = os.getenv("IMEDNET_SECURITY_KEY")
